pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "streamlit>=1.45.0",
    "tenacity>=9.1.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.2",
    "werkzeug>=3.1.3",
]
